requirements.txt
numpy
pandas
seaborn
matplotlib
streamlit
pillow
